pyproject.toml
[project]
name = "mp2-clinical-injection"
version = "0.1.0"
description = "CPIB v0.1 — Clinical Prompt Injection Benchmark pilot"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pandas>=2.0",
    "anthropic>=0.40",
    "scipy>=1.11",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "detect-secrets>=1.5",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
